Add Scores.js
